refactor: split processing.py into submodules
Replace the 2752-line monolithic handlers/processing.py with a package
of focused submodules:

- processing/audio.py      — 6 tools (analyze, QC, master, fix, reference)
- processing/sheet_music.py — 4 tools (transcribe, singles, songbook, publish)
- processing/video.py      — 2 tools (promo videos, album sampler)
- processing/mixing.py     — 3 tools (polish, analyze issues, polish pipeline)
- processing/_helpers.py   — shared dep checks, importers, title map
- processing/__init__.py   — registration + re-exports for backward compat

Uses module-level proxy with __setattr__/__getattr__ so existing test
patches (patch.object(_processing_mod, "_check_X", ...)) propagate to
the _helpers module where submodules look up functions at call time.

Zero test changes required. All 2527 tests pass.

Closes #120
